Extract KMeans TClusters into a separate file to share it with schemeshard (#19412)
