+++ ydb/core/tx/datashard/build_index/local_kmeans.cpp
@@ -47,7 +47,7 @@ using namespace NKMeans;
  *     - Output: rows annotated with cluster IDs and optional data columns
  */
 
-class TLocalKMeansScanBase: public TActor<TLocalKMeansScanBase>, public NTable::IScan {
+class TLocalKMeansScan: public TActor<TLocalKMeansScan>, public NTable::IScan {
 protected:
     using EState = NKikimrTxDataShard::EKMeansState;
 
@@ -101,7 +101,7 @@ protected:
 
     bool IsExhausted = false;
 
-    virtual TString Debug() const = 0;
+    std::unique_ptr<IClusters> Clusters;
 
 public:
     static constexpr NKikimrServices::TActivity::EType ActorActivityType()
@@ -109,11 +109,9 @@ public:
         return NKikimrServices::TActivity::LOCAL_KMEANS_SCAN_ACTOR;
     }
 
-    TLocalKMeansScanBase(ui64 tabletId, const TUserTable& table,
-                         const NKikimrTxDataShard::TEvLocalKMeansRequest& request,
-                         const TActorId& responseActorId,
-                         TAutoPtr<TEvDataShard::TEvLocalKMeansResponse>&& response,
-                         TLead&& lead)
+    TLocalKMeansScan(ui64 tabletId, const TUserTable& table, const NKikimrTxDataShard::TEvLocalKMeansRequest& request,
+        const TActorId& responseActorId, TAutoPtr<TEvDataShard::TEvLocalKMeansResponse>&& response,
+        TLead&& lead, std::unique_ptr<IClusters>&& clusters)
         : TActor{&TThis::StateWork}
         , Parent{request.GetParentFrom()}
         , Child{request.GetChild()}
@@ -129,7 +127,11 @@ public:
         , ResponseActorId{responseActorId}
         , Response{std::move(response)}
         , PrefixColumns{request.GetParentFrom() == 0 && request.GetParentTo() == 0 ? 0u : 1u}
+        , Clusters(std::move(clusters))
     {
+        LOG_I("Create " << Debug());
+        Clusters->Init(request.GetK(), request.GetNeedsRounds());
+
         const auto& embedding = request.GetEmbeddingColumn();
         const auto& data = request.GetDataColumns();
         ScanTags = MakeScanTags(table, embedding, data, EmbeddingPos, DataPos, EmbeddingTag);
@@ -163,7 +165,7 @@ public:
         auto& record = Response->Record;
         record.SetReadRows(ReadRows);
         record.SetReadBytes(ReadBytes);
-        
+
         Uploader.Finish(record, abort);
 
         if (Response->Record.GetStatus() == NKikimrIndexBuilder::DONE) {
@@ -187,6 +189,84 @@ public:
     {
         LOG_T("PageFault " << Debug());
         return EScan::Feed;
+    }
+
+    EScan Seek(TLead& lead, ui64 seq) noexcept final
+    {
+        try {
+            LOG_T("Seek " << seq << " " << Debug());
+
+            if (IsExhausted) {
+                return Uploader.CanFinish()
+                    ? EScan::Final
+                    : EScan::Sleep;
+            }
+
+            lead = Lead;
+
+            return EScan::Feed;
+        } catch (const std::exception& exc) {
+            Uploader.AddIssue(exc);
+            return EScan::Final;
+        }
+    }
+
+    EScan Feed(TArrayRef<const TCell> key, const TRow& row) noexcept final
+    {
+        try {
+            // LOG_T("Feed " << Debug());
+
+            ++ReadRows;
+            ReadBytes += CountBytes(key, row);
+
+            if (PrefixColumns && Prefix && !TCellVectorsEquals{}(Prefix.GetCells(), key.subspan(0, PrefixColumns))) {
+                if (!FinishPrefix()) {
+                    // scan current prefix rows with a new state again
+                    return EScan::Reset;
+                }
+            }
+
+            if (PrefixColumns && !Prefix) {
+                Prefix = TSerializedCellVec{key.subspan(0, PrefixColumns)};
+                auto newParent = key.at(0).template AsValue<ui64>();
+                Child += (newParent - Parent) * Clusters->GetK();
+                Parent = newParent;
+            }
+
+            if (IsFirstPrefixFeed && IsPrefixRowsValid) {
+                PrefixRows.AddRow(TSerializedCellVec{key}, TSerializedCellVec::Serialize(*row));
+                if (HasReachedLimits(PrefixRows, ScanSettings)) {
+                    PrefixRows.Clear();
+                    IsPrefixRowsValid = false;
+                }
+            }
+
+            Feed(key, *row);
+
+            return Uploader.ShouldWaitUpload() ? EScan::Sleep : EScan::Feed;
+        } catch (const std::exception& exc) {
+            Uploader.AddIssue(exc);
+            return EScan::Final;
+        }
+    }
+
+    EScan Exhausted() noexcept final
+    {
+        try {
+            LOG_T("Exhausted " << Debug());
+
+            if (!FinishPrefix()) {
+                return EScan::Reset;
+            }
+
+            IsExhausted = true;
+
+            // call Seek to wait uploads
+            return EScan::Reset;
+        } catch (const std::exception& exc) {
+            Uploader.AddIssue(exc);
+            return EScan::Final;
+        }
     }
 
 protected:
@@ -234,13 +314,9 @@ protected:
 
         Driver->Touch(EScan::Final);
     }
-};
 
-template <typename TMetric>
-class TLocalKMeansScan final : public TLocalKMeansScanBase {
-    TClusters<TMetric> Clusters;
-
-    void StartNewPrefix() {
+    void StartNewPrefix()
+    {
         State = EState::SAMPLE;
         Lead.Valid = true;
         Lead.Key = TSerializedCellVec(Prefix.GetCells()); // seek to (prefix, inf)
@@ -250,108 +326,9 @@ class TLocalKMeansScan final : public TLocalKMeansScanBase {
         IsPrefixRowsValid = true;
         PrefixRows.Clear();
         Sampler.Finish();
-        Clusters.Clear();
+        Clusters->Clear();
     }
 
-    TString Debug() const
-    {
-        return TStringBuilder() << "TLocalKMeansScan TabletId: " << TabletId << " Id: " << BuildId
-            << " State: " << State
-            << " Parent: " << Parent << " Child: " << Child
-            << " " << Sampler.Debug()
-            << " " << Clusters.Debug()
-            << " " << Uploader.Debug();
-    }
-
-public:
-    TLocalKMeansScan(ui64 tabletId, const TUserTable& table, NKikimrTxDataShard::TEvLocalKMeansRequest& request,
-        const TActorId& responseActorId, TAutoPtr<TEvDataShard::TEvLocalKMeansResponse>&& response,
-        TLead&& lead)
-        : TLocalKMeansScanBase{tabletId, table, request, responseActorId, std::move(response), std::move(lead)}
-        , Clusters(request.GetK(), request.GetSettings().vector_dimension(), request.GetNeedsRounds())
-    {
-        LOG_I("Create " << Debug());
-    }
-
-    EScan Seek(TLead& lead, ui64 seq) noexcept final
-    {
-        try {
-            LOG_T("Seek " << seq << " " << Debug());
-
-            if (IsExhausted) {
-                return Uploader.CanFinish()
-                    ? EScan::Final
-                    : EScan::Sleep;
-            }
-
-            lead = Lead;
-
-            return EScan::Feed;
-        } catch (const std::exception& exc) {
-            Uploader.AddIssue(exc);
-            return EScan::Final;
-        }
-    }
-
-    EScan Feed(TArrayRef<const TCell> key, const TRow& row) noexcept final
-    {
-        try {
-            // LOG_T("Feed " << Debug());
-
-            ++ReadRows;
-            ReadBytes += CountBytes(key, row);
-
-            if (PrefixColumns && Prefix && !TCellVectorsEquals{}(Prefix.GetCells(), key.subspan(0, PrefixColumns))) {
-                if (!FinishPrefix()) {
-                    // scan current prefix rows with a new state again
-                    return EScan::Reset;
-                }
-            }
-
-            if (PrefixColumns && !Prefix) {
-                Prefix = TSerializedCellVec{key.subspan(0, PrefixColumns)};
-                auto newParent = key.at(0).template AsValue<ui64>();
-                Child += (newParent - Parent) * Clusters.GetK();
-                Parent = newParent;
-            }
-
-            if (IsFirstPrefixFeed && IsPrefixRowsValid) {
-                PrefixRows.AddRow(TSerializedCellVec{key}, TSerializedCellVec::Serialize(*row));
-                if (HasReachedLimits(PrefixRows, ScanSettings)) {
-                    PrefixRows.Clear();
-                    IsPrefixRowsValid = false;
-                }
-            }
-
-            Feed(key, *row);
-
-            return Uploader.ShouldWaitUpload() ? EScan::Sleep : EScan::Feed;
-        } catch (const std::exception& exc) {
-            Uploader.AddIssue(exc);
-            return EScan::Final;
-        }
-    }
-
-    EScan Exhausted() noexcept final
-    {
-        try {
-            LOG_T("Exhausted " << Debug());
-
-            if (!FinishPrefix()) {
-                return EScan::Reset;
-            }
-                
-            IsExhausted = true;
-            
-            // call Seek to wait uploads
-            return EScan::Reset;
-        } catch (const std::exception& exc) {
-            Uploader.AddIssue(exc);
-            return EScan::Final;
-        }
-    }
-
-private:
     bool FinishPrefix()
     {
         if (FinishPrefixImpl()) {
@@ -360,7 +337,7 @@ private:
             return true;
         } else {
             IsFirstPrefixFeed = false;
-            
+
             if (IsPrefixRowsValid) {
                 LOG_T("FinishPrefix not finished, manually feeding " << PrefixRows.Size() << " saved rows " << Debug());
                 for (ui64 iteration = 0; ; iteration++) {
@@ -388,16 +365,17 @@ private:
     {
         if (State == EState::SAMPLE) {
             State = EState::KMEANS;
-            if (!Clusters.InitAggregatedClusters(Sampler)) {
+            if (!Clusters->SetClusters(Sampler.Finish().second)) {
                 // We don't need to do anything,
                 // because this datashard doesn't have valid embeddings for this prefix
                 return true;
             }
+            Clusters->InitAggregatedClusters();
             return false; // do KMEANS
         }
 
         if (State == EState::KMEANS) {
-            if (Clusters.RecomputeClusters()) {
+            if (Clusters->RecomputeClusters()) {
                 FormLevelRows();
                 State = UploadState;
                 return false; // do UPLOAD_*
@@ -443,7 +421,7 @@ private:
     void FeedSample(TArrayRef<const TCell> row)
     {
         const auto embedding = row.at(EmbeddingPos).AsRef();
-        if (!IsExpectedSize<typename TMetric::TCoord_>(embedding, Dimensions)) {
+        if (!Clusters->IsExpectedSize(embedding)) {
             return;
         }
 
@@ -454,35 +432,35 @@ private:
 
     void FeedKMeans(TArrayRef<const TCell> row) noexcept
     {
-        if (auto pos = Clusters.FindCluster(row, EmbeddingPos); pos) {
-            Clusters.AggregateToCluster(*pos, row.at(EmbeddingPos).Data());
+        if (auto pos = Clusters->FindCluster(row, EmbeddingPos); pos) {
+            Clusters->AggregateToCluster(*pos, row.at(EmbeddingPos).Data());
         }
     }
 
     void FeedMainToBuild(TArrayRef<const TCell> key, TArrayRef<const TCell> row) noexcept
     {
-        if (auto pos = Clusters.FindCluster(row, EmbeddingPos); pos) {
+        if (auto pos = Clusters->FindCluster(row, EmbeddingPos); pos) {
             AddRowToData(*OutputBuf, Child + *pos, key, row, key, false);
         }
     }
 
     void FeedMainToPosting(TArrayRef<const TCell> key, TArrayRef<const TCell> row) noexcept
     {
-        if (auto pos = Clusters.FindCluster(row, EmbeddingPos); pos) {
+        if (auto pos = Clusters->FindCluster(row, EmbeddingPos); pos) {
             AddRowToData(*OutputBuf, Child + *pos, key, row.Slice(DataPos), key, true);
         }
     }
 
     void FeedBuildToBuild(TArrayRef<const TCell> key, TArrayRef<const TCell> row) noexcept
     {
-        if (auto pos = Clusters.FindCluster(row, EmbeddingPos); pos) {
+        if (auto pos = Clusters->FindCluster(row, EmbeddingPos); pos) {
             AddRowToData(*OutputBuf, Child + *pos, key.Slice(1), row, key, false);
         }
     }
 
     void FeedBuildToPosting(TArrayRef<const TCell> key, TArrayRef<const TCell> row) noexcept
     {
-        if (auto pos = Clusters.FindCluster(row, EmbeddingPos); pos) {
+        if (auto pos = Clusters->FindCluster(row, EmbeddingPos); pos) {
             AddRowToData(*OutputBuf, Child + *pos, key.Slice(1), row.Slice(DataPos), key, true);
         }
     }
@@ -492,10 +470,20 @@ private:
         const bool isPostingLevel = UploadState == NKikimrTxDataShard::UPLOAD_MAIN_TO_POSTING
             || UploadState == NKikimrTxDataShard::UPLOAD_BUILD_TO_POSTING;
 
-        for (NTable::TPos pos = 0; const auto& row : Clusters.GetClusters()) {
+        for (NTable::TPos pos = 0; const auto& row : Clusters->GetClusters()) {
             AddRowToLevel(*LevelBuf, Parent, Child + pos, row, isPostingLevel);
             ++pos;
         }
+    }
+
+    TString Debug() const
+    {
+        return TStringBuilder() << "TLocalKMeansScan TabletId: " << TabletId << " Id: " << BuildId
+            << " State: " << State
+            << " Parent: " << Parent << " Child: " << Child
+            << " " << Sampler.Debug()
+            << " " << Clusters->Debug()
+            << " " << Uploader.Debug();
     }
 };
 
@@ -661,19 +649,18 @@ void TDataShard::HandleSafe(TEvDataShard::TEvLocalKMeansRequest::TPtr& ev, const
         }
 
         // 3. Validating vector index settings
-        TAutoPtr<NTable::IScan> scan;
-        auto createScan = [&]<typename T> {
-            scan = new TLocalKMeansScan<T>{
-                TabletID(), userTable, request, ev->Sender, std::move(response),
-                std::move(lead)
-            };
-        };
-        MakeScan(request, createScan, badRequest);
-        if (!scan) {
+        TString error;
+        auto clusters = NKikimr::NKMeans::CreateClusters(request.GetSettings(), error);
+        if (!clusters) {
+            badRequest(error);
             auto sent = trySendBadRequest();
             Y_ENSURE(sent);
             return;
         }
+        TAutoPtr<NTable::IScan> scan = new TLocalKMeansScan(
+            TabletID(), userTable, request, ev->Sender, std::move(response),
+            std::move(lead), std::move(clusters)
+        );
 
         StartScan(this, std::move(scan), id, seqNo, rowVersion, userTable.LocalTid);
     } catch (const std::exception& exc) {
