+++ ydb/core/tx/datashard/build_index/reshuffle_kmeans.cpp
@@ -46,7 +46,7 @@ using namespace NKMeans;
  *   - The row and any specified data columns are written to the output table
  */
 
-class TReshuffleKMeansScanBase: public TActor<TReshuffleKMeansScanBase>, public NTable::IScan {
+class TReshuffleKMeansScan: public TActor<TReshuffleKMeansScan>, public NTable::IScan {
 protected:
     using EState = NKikimrTxDataShard::EKMeansState;
 
@@ -66,7 +66,7 @@ protected:
     ui64 ReadBytes = 0;
 
     TBatchRowsUploader Uploader;
-    
+
     TBufferData* OutputBuf = nullptr;
 
     const ui32 Dimensions = 0;
@@ -86,7 +86,7 @@ protected:
 
     bool IsExhausted = false;
 
-    virtual TString Debug() const = 0;
+    std::unique_ptr<IClusters> Clusters;
 
 public:
     static constexpr NKikimrServices::TActivity::EType ActorActivityType()
@@ -94,23 +94,27 @@ public:
         return NKikimrServices::TActivity::RESHUFFLE_KMEANS_SCAN_ACTOR;
     }
 
-    TReshuffleKMeansScanBase(ui64 tabletId, const TUserTable& table, TLead&& lead,
-                             const NKikimrTxDataShard::TEvReshuffleKMeansRequest& request,
-                             const TActorId& responseActorId,
-                             TAutoPtr<TEvDataShard::TEvReshuffleKMeansResponse>&& response)
-        : TActor{&TThis::StateWork}
-        , Parent{request.GetParent()}
-        , Child{request.GetChild()}
-        , UploadState{request.GetUpload()}
-        , Lead{std::move(lead)}
+    TReshuffleKMeansScan(ui64 tabletId, const TUserTable& table, TLead&& lead,
+        const NKikimrTxDataShard::TEvReshuffleKMeansRequest& request,
+        const TActorId& responseActorId, TAutoPtr<TEvDataShard::TEvReshuffleKMeansResponse>&& response,
+        std::unique_ptr<IClusters>&& clusters)
+        : TActor(&TThis::StateWork)
+        , Parent(request.GetParent())
+        , Child(request.GetChild())
+        , UploadState(request.GetUpload())
+        , Lead(std::move(lead))
         , TabletId(tabletId)
-        , BuildId{request.GetId()}
+        , BuildId(request.GetId())
         , Uploader(request.GetScanSettings())
         , Dimensions(request.GetSettings().vector_dimension())
         , ScanSettings(request.GetScanSettings())
-        , ResponseActorId{responseActorId}
-        , Response{std::move(response)}
+        , ResponseActorId(responseActorId)
+        , Response(std::move(response))
+        , Clusters(std::move(clusters))
     {
+        LOG_I("Create " << Debug());
+        Clusters->SetClusters(TVector<TString>{request.GetClusters().begin(), request.GetClusters().end()});
+
         const auto& embedding = request.GetEmbeddingColumn();
         const auto& data = request.GetDataColumns();
         NTable::TTag embeddingTag;
@@ -135,7 +139,7 @@ public:
         auto& record = Response->Record;
         record.SetReadRows(ReadRows);
         record.SetReadBytes(ReadBytes);
-        
+
         Uploader.Finish(record, abort);
 
         if (Response->Record.GetStatus() == NKikimrIndexBuilder::DONE) {
@@ -159,74 +163,6 @@ public:
     {
         LOG_T("PageFault " << Debug());
         return EScan::Feed;
-    }
-
-protected:
-    STFUNC(StateWork)
-    {
-        switch (ev->GetTypeRewrite()) {
-            HFunc(TEvTxUserProxy::TEvUploadRowsResponse, Handle);
-            CFunc(TEvents::TSystem::Wakeup, HandleWakeup);
-            default:
-                LOG_E("StateWork unexpected event type: " << ev->GetTypeRewrite() 
-                    << " event: " << ev->ToString() << " " << Debug());
-        }
-    }
-
-    void HandleWakeup(const NActors::TActorContext& /*ctx*/)
-    {
-        LOG_D("Retry upload " << Debug());
-
-        Uploader.RetryUpload();
-    }
-
-    void Handle(TEvTxUserProxy::TEvUploadRowsResponse::TPtr& ev, const TActorContext& ctx)
-    {
-        LOG_D("Handle TEvUploadRowsResponse " << Debug()
-            << " ev->Sender: " << ev->Sender.ToString());
-
-        if (!Driver) {
-            return;
-        }
-
-        Uploader.Handle(ev);
-
-        if (Uploader.GetUploadStatus().IsSuccess()) {
-            Driver->Touch(EScan::Feed);
-            return;
-        }
-
-        if (auto retryAfter = Uploader.GetRetryAfter(); retryAfter) {
-            LOG_N("Got retriable error, " << Debug() << " " << Uploader.GetUploadStatus().ToString());
-            ctx.Schedule(*retryAfter, new TEvents::TEvWakeup());
-            return;
-        }
-
-        LOG_N("Got error, abort scan, " << Debug() << " " << Uploader.GetUploadStatus().ToString());
-
-        Driver->Touch(EScan::Final);
-    }
-};
-
-template <typename TMetric>
-class TReshuffleKMeansScan final : public TReshuffleKMeansScanBase {
-    TClusters<TMetric> Clusters;
-
-    TString Debug() const
-    {
-        return TStringBuilder() << "TReshuffleKMeansScan TabletId: " << TabletId << " Id: " << BuildId
-            << " Parent: " << Parent << " Child: " << Child
-            << " " << Clusters.Debug()
-            << " " << Uploader.Debug();
-    }
-
-public:
-    TReshuffleKMeansScan(ui64 tabletId, const TUserTable& table, TLead&& lead, NKikimrTxDataShard::TEvReshuffleKMeansRequest& request,
-                         const TActorId& responseActorId, TAutoPtr<TEvDataShard::TEvReshuffleKMeansResponse>&& response)
-        : TReshuffleKMeansScanBase{tabletId, table, std::move(lead), request, responseActorId, std::move(response)}
-        , Clusters(TVector<TString>{request.GetClusters().begin(), request.GetClusters().end()}, request.GetSettings().vector_dimension())
-    {
-        LOG_I("Create " << Debug());
     }
 
     EScan Seek(TLead& lead, ui64 seq) noexcept final
@@ -272,7 +208,7 @@ public:
             LOG_T("Exhausted " << Debug());
 
             IsExhausted = true;
-            
+
             // call Seek to wait uploads
             return EScan::Reset;
         } catch (const std::exception& exc) {
@@ -281,7 +217,60 @@ public:
         }
     }
 
-private:
+protected:
+    STFUNC(StateWork)
+    {
+        switch (ev->GetTypeRewrite()) {
+            HFunc(TEvTxUserProxy::TEvUploadRowsResponse, Handle);
+            CFunc(TEvents::TSystem::Wakeup, HandleWakeup);
+            default:
+                LOG_E("StateWork unexpected event type: " << ev->GetTypeRewrite()
+                    << " event: " << ev->ToString() << " " << Debug());
+        }
+    }
+
+    void HandleWakeup(const NActors::TActorContext& /*ctx*/)
+    {
+        LOG_D("Retry upload " << Debug());
+
+        Uploader.RetryUpload();
+    }
+
+    void Handle(TEvTxUserProxy::TEvUploadRowsResponse::TPtr& ev, const TActorContext& ctx)
+    {
+        LOG_D("Handle TEvUploadRowsResponse " << Debug()
+            << " ev->Sender: " << ev->Sender.ToString());
+
+        if (!Driver) {
+            return;
+        }
+
+        Uploader.Handle(ev);
+
+        if (Uploader.GetUploadStatus().IsSuccess()) {
+            Driver->Touch(EScan::Feed);
+            return;
+        }
+
+        if (auto retryAfter = Uploader.GetRetryAfter(); retryAfter) {
+            LOG_N("Got retriable error, " << Debug() << " " << Uploader.GetUploadStatus().ToString());
+            ctx.Schedule(*retryAfter, new TEvents::TEvWakeup());
+            return;
+        }
+
+        LOG_N("Got error, abort scan, " << Debug() << " " << Uploader.GetUploadStatus().ToString());
+
+        Driver->Touch(EScan::Final);
+    }
+
+    TString Debug() const
+    {
+        return TStringBuilder() << "TReshuffleKMeansScan TabletId: " << TabletId << " Id: " << BuildId
+            << " Parent: " << Parent << " Child: " << Child
+            << " " << Clusters->Debug()
+            << " " << Uploader.Debug();
+    }
+
     void Feed(TArrayRef<const TCell> key, TArrayRef<const TCell> row)
     {
         switch (UploadState) {
@@ -304,28 +293,28 @@ private:
 
     void FeedMainToBuild(TArrayRef<const TCell> key, TArrayRef<const TCell> row) noexcept
     {
-        if (auto pos = Clusters.FindCluster(row, EmbeddingPos); pos) {
+        if (auto pos = Clusters->FindCluster(row, EmbeddingPos); pos) {
             AddRowToData(*OutputBuf, Child + *pos, key, row, key, false);
         }
     }
 
     void FeedMainToPosting(TArrayRef<const TCell> key, TArrayRef<const TCell> row) noexcept
     {
-        if (auto pos = Clusters.FindCluster(row, EmbeddingPos); pos) {
+        if (auto pos = Clusters->FindCluster(row, EmbeddingPos); pos) {
             AddRowToData(*OutputBuf, Child + *pos, key, row.Slice(DataPos), key, true);
         }
     }
 
     void FeedBuildToBuild(TArrayRef<const TCell> key, TArrayRef<const TCell> row) noexcept
     {
-        if (auto pos = Clusters.FindCluster(row, EmbeddingPos); pos) {
+        if (auto pos = Clusters->FindCluster(row, EmbeddingPos); pos) {
             AddRowToData(*OutputBuf, Child + *pos, key.Slice(1), row, key, false);
         }
     }
 
     void FeedBuildToPosting(TArrayRef<const TCell> key, TArrayRef<const TCell> row) noexcept
     {
-        if (auto pos = Clusters.FindCluster(row, EmbeddingPos); pos) {
+        if (auto pos = Clusters->FindCluster(row, EmbeddingPos); pos) {
             AddRowToData(*OutputBuf, Child + *pos, key.Slice(1), row.Slice(DataPos), key, true);
         }
     }
@@ -374,7 +363,7 @@ void TDataShard::HandleSafe(TEvDataShard::TEvReshuffleKMeansRequest::TPtr& ev, c
         response->Record.SetRequestSeqNoGeneration(seqNo.Generation);
         response->Record.SetRequestSeqNoRound(seqNo.Round);
 
-        LOG_N("Starting TReshuffleKMeansScan TabletId: " << TabletID() 
+        LOG_N("Starting TReshuffleKMeansScan TabletId: " << TabletID()
             << " " << ToShortDebugString(request)
             << " row version " << rowVersion);
 
@@ -443,13 +432,13 @@ void TDataShard::HandleSafe(TEvDataShard::TEvReshuffleKMeansRequest::TPtr& ev, c
         const auto parent = request.GetParent();
         NTable::TLead lead;
         if (parent == 0) {
-            if (request.GetUpload() == NKikimrTxDataShard::UPLOAD_BUILD_TO_BUILD 
+            if (request.GetUpload() == NKikimrTxDataShard::UPLOAD_BUILD_TO_BUILD
                 || request.GetUpload() == NKikimrTxDataShard::UPLOAD_BUILD_TO_POSTING)
             {
                 badRequest("Wrong upload for zero parent");
             }
             lead.To({}, NTable::ESeek::Lower);
-        } else if (request.GetUpload() == NKikimrTxDataShard::UPLOAD_MAIN_TO_BUILD 
+        } else if (request.GetUpload() == NKikimrTxDataShard::UPLOAD_MAIN_TO_BUILD
             || request.GetUpload() == NKikimrTxDataShard::UPLOAD_MAIN_TO_POSTING)
         {
             badRequest("Wrong upload for non-zero parent");
@@ -481,18 +470,17 @@ void TDataShard::HandleSafe(TEvDataShard::TEvReshuffleKMeansRequest::TPtr& ev, c
         }
 
         // 3. Validating vector index settings
-        TAutoPtr<NTable::IScan> scan;
-        auto createScan = [&]<typename T> {
-            scan = new TReshuffleKMeansScan<T>{
-                TabletID(), userTable, std::move(lead), request, ev->Sender, std::move(response),
-            };
-        };
-        MakeScan(request, createScan, badRequest);
-        if (!scan) {
+        TString error;
+        auto clusters = NKikimr::NKMeans::CreateClusters(request.GetSettings(), error);
+        if (!clusters) {
+            badRequest(error);
             auto sent = trySendBadRequest();
             Y_ENSURE(sent);
             return;
         }
+        TAutoPtr<NTable::IScan> scan = new TReshuffleKMeansScan(
+            TabletID(), userTable, std::move(lead), request, ev->Sender, std::move(response), std::move(clusters)
+        );
 
         StartScan(this, std::move(scan), id, seqNo, rowVersion, userTable.LocalTid);
     } catch (const std::exception& exc) {
